requirements.txt
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
Django==1.11.9
django-appconf==1.0.2
django-compressor==2.2
django-extra-views==0.6.4
django-haystack==2.6.1
django-localflavor==2.0
django-oscar==1.5.1
django-oscar-accounts==0.3
django-phonenumber-field==1.3.0
django-tables2==1.17.1
django-treebeard==4.2.0
django-widget-tweaks==1.4.1
factory-boy==2.9.2
Faker==0.8.10
gunicorn==19.7.1
idna==2.6
mock==2.0.0
pbr==3.1.1
phonenumberslite==8.8.9
Pillow==5.0.0
purl==1.3.1
pycountry==17.9.23
python-dateutil==2.1
pytz==2017.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.1
Unidecode==0.4.21
urllib3==1.22
whitenoise==3.3.1
